chore: revisa algumas configurações

+++ config/functions.php
@@ -8,7 +8,6 @@ if (! function_exists('toArray')) {
     /**
      * @template T of array-key
      * @template U
-     * @param mixed $value
      * @param array<T, U> $default
      * @return array<T, U>
      */
@@ -53,7 +52,6 @@ if (! function_exists('extractArray')) {
      * @template T
      * @template U
      * @param array<string, array<T, U>> $array
-     * @param string $property
      * @param array<T, U> $default
      * @return array<T, U>
      */
@@ -70,9 +68,6 @@ if (! function_exists('extractArray')) {
 if (! function_exists('extractString')) {
     /**
      * @param array<string, mixed> $array
-     * @param string $property
-     * @param string $default
-     * @return string
      */
     function extractString(array $array, string $property, string $default = ''): string
     {
@@ -84,9 +79,6 @@ if (! function_exists('extractString')) {
 if (! function_exists('extractInt')) {
     /**
      * @param array<string, mixed> $array
-     * @param string $property
-     * @param int $default
-     * @return int
      */
     function extractInt(array $array, string $property, int $default = 0): int
     {
@@ -98,9 +90,6 @@ if (! function_exists('extractInt')) {
 if (! function_exists('extractBool')) {
     /**
      * @param array<string, mixed> $array
-     * @param string $property
-     * @param bool $default
-     * @return bool
      */
     function extractBool(array $array, string $property, bool $default = false): bool
     {
@@ -112,11 +101,8 @@ if (! function_exists('extractBool')) {
 if (! function_exists('extractNumeric')) {
     /**
      * @param array<string, mixed> $array
-     * @param string $property
-     * @param int|float $default
-     * @return float
      */
-    function extractNumeric(array $array, string $property, int|float $default = 0): float
+    function extractNumeric(array $array, string $property, float|int $default = 0): float
     {
         $numeric = $array[$property] ?? $default;
         return (float) (is_numeric($numeric) ? $numeric : $default);
